feat: extract scoring.py module from prospect.py
Issue #11: Centralized scoring logic for reuse across commands.

- Create pf_scout/scoring.py with:
  - TECH_KEYWORDS and QUANT_KEYWORDS constants
  - DEFAULT_DIMENSIONS list
  - apply_keyword_heuristics() for keyword-based scoring
  - score_dimension() for single dimension scoring
  - score_contact() for full contact scoring with tier
  - Individual scorers: score_technical_depth, score_forecasting,
    score_operational_reliability, score_engagement_consistency
  - Utility functions: get_text_blob, evidence_sentence, infer_role

- Refactor prospect.py to import from scoring module
- Maintain backward compatibility via re-exports
