+++ pf_scout/commands/prospect.py
@@ -9,199 +9,46 @@ import requests
 import yaml
 
 from ..db import get_connection
+from ..scoring import (
+    DEFAULT_DIMENSIONS,
+    QUANT_KEYWORDS,
+    TECH_KEYWORDS,
+    evidence_sentence,
+    get_text_blob,
+    infer_role,
+    score_contact,
+    score_engagement_consistency,
+    score_forecasting,
+    score_operational_reliability,
+    score_technical_depth,
+)
 
 
-# ---------------------------------------------------------------------------
-# Default scoring dimensions (used when no rubric YAML provided)
-# ---------------------------------------------------------------------------
-
-TECH_KEYWORDS = [
-    "infrastructure", "devops", "backend", "engineering", "smart contract",
-    "blockchain", "solidity", "rust", "python", "go", "typescript",
-    "kubernetes", "docker", "cloud", "aws", "api", "protocol", "security",
-    "cryptography", "zk", "evm", "validator", "rpc",
+# Re-export for backward compatibility with existing imports (e.g., tests)
+__all__ = [
+    "DEFAULT_DIMENSIONS",
+    "TECH_KEYWORDS",
+    "QUANT_KEYWORDS",
+    "score_technical_depth",
+    "score_forecasting",
+    "score_operational_reliability",
+    "score_engagement_consistency",
+    "score_row",
+    "generate_document",
 ]
-
-QUANT_KEYWORDS = [
-    "quant", "quantitative", "machine learning", "ml", "statistics",
-    "trading", "signal", "forecasting", "data science", "analytics",
-    "on-chain", "onchain", "backtesting", "risk management", "portfolio",
-    "financial modeling", "time series", "alpha", "macro", "research",
-    "modeling", "prediction", "probability",
-]
-
-DEFAULT_DIMENSIONS = [
-    {"key": "technical_depth", "label": "Technical Depth", "weight": 1},
-    {"key": "forecasting", "label": "Forecasting / Quantitative Potential", "weight": 1},
-    {"key": "operational_reliability", "label": "Operational Reliability", "weight": 1},
-    {"key": "engagement_consistency", "label": "Engagement Consistency", "weight": 1},
-]
-
-
-def _text_blob(row: dict) -> str:
-    """Concatenate all text fields from a leaderboard row for keyword search."""
-    parts = []
-    if row.get("summary"):
-        parts.append(str(row["summary"]))
-    for cap in row.get("capabilities") or []:
-        if isinstance(cap, dict):
-            parts.append(" ".join(str(v) for v in cap.values()))
-        else:
-            parts.append(str(cap))
-    for ek in row.get("expert_knowledge") or []:
-        if isinstance(ek, dict):
-            parts.append(" ".join(str(v) for v in ek.values()))
-        else:
-            parts.append(str(ek))
-    return " ".join(parts).lower()
-
-
-def _count_keyword_hits(text: str, keywords: list[str]) -> int:
-    """Count how many distinct keywords appear in text."""
-    return sum(1 for kw in keywords if kw in text)
-
-
-def _hits_to_base(hits: int) -> int:
-    if hits == 0:
-        return 1
-    if hits == 1:
-        return 2
-    if hits == 2:
-        return 3
-    if hits <= 4:
-        return 4
-    return 5
-
-
-def score_technical_depth(row: dict) -> int:
-    text = _text_blob(row)
-    base = _hits_to_base(_count_keyword_hits(text, TECH_KEYWORDS))
-    if (row.get("sybil_score") or 0) >= 85:
-        base += 1
-    return min(base, 5)
-
-
-def score_forecasting(row: dict) -> int:
-    text = _text_blob(row)
-    base = _hits_to_base(_count_keyword_hits(text, QUANT_KEYWORDS))
-    if (row.get("alignment_score") or 0) >= 90:
-        base += 1
-    if (row.get("monthly_rewards") or 0) > 500000:
-        base += 1
-    return min(base, 5)
-
-
-def score_operational_reliability(row: dict) -> int:
-    lsm = row.get("leaderboard_score_month") or 0
-    if lsm < 15:
-        base = 1
-    elif lsm < 40:
-        base = 2
-    elif lsm < 60:
-        base = 3
-    elif lsm < 80:
-        base = 4
-    else:
-        base = 5
-    if (row.get("monthly_tasks") or 0) >= 30:
-        base += 1
-    monthly = row.get("monthly_rewards") or 0
-    weekly = row.get("weekly_rewards") or 0
-    if monthly > 0:
-        ratio = weekly / monthly
-        if 0.2 <= ratio <= 0.35:
-            base += 1
-    return min(base, 5)
-
-
-def score_engagement_consistency(row: dict) -> int:
-    weekly = row.get("weekly_rewards") or 0
-    if weekly <= 0:
-        base = 1
-    elif weekly <= 50000:
-        base = 2
-    elif weekly <= 150000:
-        base = 3
-    elif weekly <= 300000:
-        base = 4
-    else:
-        base = 5
-    if (row.get("leaderboard_score_week") or 0) > 50:
-        base += 1
-    return min(base, 5)
-
-
-SCORERS = {
-    "technical_depth": score_technical_depth,
-    "forecasting": score_forecasting,
-    "operational_reliability": score_operational_reliability,
-    "engagement_consistency": score_engagement_consistency,
-}
 
 
 def score_row(row: dict, dimensions: list[dict]) -> dict:
-    """Score a single leaderboard row against the given dimensions."""
-    scores = {}
-    for dim in dimensions:
-        key = dim["key"]
-        scorer = SCORERS.get(key)
-        scores[key] = scorer(row) if scorer else 1
-    composite = sum(scores.values())
-    max_possible = len(dimensions) * 5
-    pct = composite / max_possible if max_possible else 0
-    if pct >= 0.8:
-        tier = "🔴 Top Tier"
-    elif pct >= 0.6:
-        tier = "🟡 Mid Tier"
-    else:
-        tier = "⚪ Speculative"
-    return {
-        "scores": scores,
-        "composite": composite,
-        "max": max_possible,
-        "pct": pct,
-        "tier": tier,
-    }
-
-
-def _evidence_sentence(row: dict, dim_key: str) -> str:
-    """Generate a brief evidence sentence for a dimension score."""
-    text = _text_blob(row)
-    if dim_key == "technical_depth":
-        hits = [kw for kw in TECH_KEYWORDS if kw in text]
-        if hits:
-            return f"Keywords matched: {', '.join(hits[:5])}."
-        return "No technical keywords found in profile."
-    if dim_key == "forecasting":
-        hits = [kw for kw in QUANT_KEYWORDS if kw in text]
-        if hits:
-            return f"Keywords matched: {', '.join(hits[:5])}."
-        return "No quantitative keywords found in profile."
-    if dim_key == "operational_reliability":
-        lsm = row.get("leaderboard_score_month") or 0
-        mt = row.get("monthly_tasks") or 0
-        return f"Leaderboard score (month): {lsm}, monthly tasks: {mt}."
-    if dim_key == "engagement_consistency":
-        wr = row.get("weekly_rewards") or 0
-        lsw = row.get("leaderboard_score_week") or 0
-        return f"Weekly rewards: {wr:,.0f}, leaderboard score (week): {lsw}."
-    return "No evidence available."
-
-
-def _infer_role(row: dict) -> str:
-    text = _text_blob(row)
-    if any(kw in text for kw in ["trading", "quant", "alpha", "signal"]):
-        return "Signal / Quant"
-    if any(kw in text for kw in ["infrastructure", "devops", "kubernetes", "docker"]):
-        return "Infrastructure"
-    if any(kw in text for kw in ["smart contract", "solidity", "evm", "blockchain"]):
-        return "Protocol Engineer"
-    if any(kw in text for kw in ["research", "analytics", "data science"]):
-        return "Researcher"
-    return "Contributor"
+    """Score a single leaderboard row against the given dimensions.
+    
+    This is a thin wrapper around scoring.score_contact for backward
+    compatibility with existing code and tests.
+    """
+    return score_contact(row, dimensions)
 
 
 def _fetch_leaderboard(jwt: str, base_url: str) -> list[dict]:
+    """Fetch live leaderboard data from the Post Fiat API."""
     resp = requests.get(
         f"{base_url}/api/leaderboard",
         headers={
@@ -215,6 +62,7 @@ def _fetch_leaderboard(jwt: str, base_url: str) -> list[dict]:
 
 
 def _load_from_db(db_path: str) -> list[dict]:
+    """Load leaderboard data from local database snapshots."""
     conn = get_connection(db_path)
     rows = conn.execute(
         "SELECT payload FROM signals WHERE signal_type = 'postfiat/leaderboard' "
@@ -234,6 +82,7 @@ def _load_from_db(db_path: str) -> list[dict]:
 
 
 def _load_rubric(rubric_path: str) -> tuple[str, list[dict]]:
+    """Load and parse a rubric YAML file."""
     with open(rubric_path) as f:
         rubric = yaml.safe_load(f)
     name = rubric.get("name", Path(rubric_path).stem)
@@ -255,7 +104,18 @@ def generate_document(
     title: str,
     min_composite: int,
 ) -> str:
-    """Score all rows and generate the markdown pipeline document."""
+    """Score all rows and generate the markdown pipeline document.
+    
+    Args:
+        rows: List of leaderboard row dictionaries.
+        dimensions: List of dimension dicts from rubric.
+        rubric_name: Name of the rubric for display.
+        title: Document title prefix.
+        min_composite: Minimum composite score to include.
+    
+    Returns:
+        Markdown document string.
+    """
     today = datetime.utcnow().strftime("%Y-%m-%d")
     scored = []
     for row in rows:
@@ -264,7 +124,7 @@ def generate_document(
             continue
         label = row.get("summary") or row.get("wallet_address", "unknown")
         wallet = row.get("wallet_address", "")
-        role = _infer_role(row)
+        role = infer_role(row)
         scored.append({
             "label": label,
             "wallet": wallet,
@@ -342,7 +202,7 @@ def generate_document(
 
         for d in dimensions:
             score_val = s["scores"].get(d["key"], "-")
-            evidence = _evidence_sentence(s["row"], d["key"])
+            evidence = evidence_sentence(s["row"], d["key"])
             lines.append(f"**{d['label']} ({score_val}/5):** {evidence}\n")
 
         # Assessment
